FPGAMT-7 Move packages to com.lintyservices FPGAMT-6 Compile with JDK 11 FPGAMT-5 Update metadata FPGAMT-4 Set parent POM to org.sonarsource.parent:parent
