+++ sonar-fpga-metrics-plugin/src/test/java/com/lintyservices/sonar/plugins/fpgametrics/ImportersTest.java
@@ -1,0 +1,78 @@
+/*
+ * SonarQube Linty FPGA Metrics :: Plugin
+ * Copyright (C) 2020-2020 Linty Services
+ * mailto:contact@linty-services.com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package com.lintyservices.sonar.plugins.fpgametrics;
+
+
+import com.lintyservices.sonar.plugins.fpgametrics.measures.MeasuresImporter;
+import com.lintyservices.sonar.plugins.fpgametrics.measures.MetricsImporter;
+import org.junit.Test;
+import org.sonar.api.batch.sensor.internal.SensorContextTester;
+import org.sonar.api.batch.sensor.measure.Measure;
+import org.sonar.api.internal.PluginContextImpl;
+import org.sonar.api.internal.SonarRuntimeImpl;
+import org.sonar.api.measures.Metric;
+import org.sonar.api.measures.Metric.ValueType;
+import org.sonar.api.utils.Version;
+
+import java.io.File;
+import java.util.List;
+
+import static org.junit.Assert.assertEquals;
+
+
+public class ImportersTest {
+
+
+  @Test
+  public void execute_WithMatchingPattern_ExpectValueRegistered() throws Exception {
+
+    //Instantiating MetricsPlugin class
+
+    MetricsPlugin metricsPlugin = new MetricsPlugin();
+    metricsPlugin.define((new PluginContextImpl.Builder().setSonarRuntime(SonarRuntimeImpl.forSonarLint(Version.parse("1.1.1")))).build());
+
+
+    //Testing MetricsImporter
+
+    MetricsImporter metricsImporter = new MetricsImporter();
+    List<Metric> metricsList = metricsImporter.getMetrics();
+    MetricsImporter.getJsonMetrics();
+    assertEquals(116, MetricsImporter.getMetricsResult().size());
+    assertEquals("NX logs remarks count", MetricsImporter.getMetricsResult().get(0).getName());
+    assertEquals(ValueType.INT, MetricsImporter.getMetricsResult().get(1).getType());
+    assertEquals(false, MetricsImporter.getMetricsResult().get(2).getQualitative());
+
+
+    //Testing MeasuresImporter
+
+    final SensorContextTester contextTester = SensorContextTester.create(new File("src\\test\\files\\ctx")); //Blank file only used for context simulation
+    MeasuresImporter measuresImporter = new MeasuresImporter(metricsList, "src\\test\\files\\");
+    measuresImporter.execute(contextTester);
+    final Measure<Integer> intMeasure = contextTester.measure(contextTester.module().key(), "NX_Log_Remarks");
+    final Measure<Double> floatMeasure = contextTester.measure(contextTester.module().key(), "NX_CLK1_Max_Delay");
+    assertEquals((Integer) 1, intMeasure.value());
+    assertEquals((Double) 54.385, floatMeasure.value());
+
+    measuresImporter = new MeasuresImporter(null, "src\\test\\files_folder_does_not_exist\\");
+    measuresImporter.execute(contextTester);
+  }
+
+
+}
